+++ resources/types.go
@@ -34,6 +34,7 @@ type DeploymentValues struct {
 	Cronjobs            []Cronjob
 	ConfigMaps          map[string]map[string]string
 	ServiceMonitor      *schema.ServiceMonitor
+	ServiceType         corev1.ServiceType
 
 	Annotations    map[string]string
 	PodAnnotations map[string]string