+++ schema/schema_validations.go
@@ -10,12 +10,32 @@ func CustomValidations(values InputValues) error {
 		return err
 	}
 
+	// 2. NodePorts (if specified) must be between 30000 and 32767
+	if err := validateNodePortRange(values); err != nil {
+		return err
+	}
 	return nil
 }
 
 func validateKindValue(values InputValues) error {
 	if values.Kind != nil && *values.Kind != "Deployment" && *values.Kind != "StatefulSet" {
 		return fmt.Errorf("invalid kind %v", *values.Kind)
+	}
+	return nil
+}
+
+func validateNodePortRange(values InputValues) error {
+	for _, port := range values.Ports {
+		if port.NodePort != nil && (*port.NodePort < 30000 || *port.NodePort > 32767) {
+			return fmt.Errorf("node port on port %d must be between 30000 and 32767", port.Port)
+		}
+	}
+	for name, sidecar := range values.Sidecars {
+		for _, port := range sidecar.Ports {
+			if port.NodePort != nil && (*port.NodePort < 30000 || *port.NodePort > 32767) {
+				return fmt.Errorf("node port on port %d of sidecar %s must be between 30000 and 32767", port.Port, name)
+			}
+		}
 	}
 	return nil
 }